LayoutTests/http/tests/inspector/network/har-content-expected.txt

CONSOLE MESSAGE: line 6: XHR loaded: 1
Tests conversion of Inspector's resource representation into HAR format.

initiator.css: .image-background {
    background-image: url("resource.php?type=image&random=1&size=200");
}

.image-background-2 {
    background-image: url("resource.php?type=image&random=1&size=300");
}
